Amazon/requirements.txt
Flask
beautifulsoup4
requests
psycopg2-binary
html5lib
html.parser
lxml